# Back/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "FreeLib Aplication REST-API schema",
        "title": "Goland FreeLib API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/login": {
            "post": {
                "description": "Авторизовать пользователя в системе и получить JWT токен",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Авторизовать пользователя",
                "parameters": [
                    {
                        "description": "AuthorizationUser тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.AuthorizationUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешная авторизация",
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.AuthorizationUserResponce"
                        }
                    },
                    "400": {
                        "description": "BadRequest",
                        "schema": {
                            "$ref": "#/definitions/github_com_Eternity8c_FreeLib_internal_core_transport_http_responce.ErrorResponce"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_Eternity8c_FreeLib_internal_core_transport_http_responce.ErrorResponce"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Eternity8c_FreeLib_internal_core_transport_http_responce.ErrorResponce"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Создать нового пользователя в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Создать пользователя",
                "parameters": [
                    {
                        "description": "CreateUser тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешно созданный пользователь",
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.CreateUserResponce"
                        }
                    },
                    "400": {
                        "description": "BadRequest",
                        "schema": {
                            "$ref": "#/definitions/github_com_Eternity8c_FreeLib_internal_core_transport_http_responce.ErrorResponce"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_Eternity8c_FreeLib_internal_core_transport_http_responce.ErrorResponce"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_Eternity8c_FreeLib_internal_core_transport_http_responce.ErrorResponce": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_features_users_transport_http.AuthorizationUserRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "password1234"
                }
            }
        },
        "internal_features_users_transport_http.AuthorizationUserResponce": {
            "type": "object",
            "properties": {
                "jwt_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "internal_features_users_transport_http.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "password1234"
                }
            }
        },
        "internal_features_users_transport_http.CreateUserResponce": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@gmail.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}